+++ ArcdpsLogManager/Logs/Updates/LogDataUpdater.cs
@@ -190,6 +190,9 @@ namespace GW2Scratch.ArcdpsLogManager.Logs.Updates
 				"Fix success detection for Cosmic Observatory"),
 			// When adding a new update, you need to increase the revision (last value) of the version in the .csproj file
 			// unless the version changes more significantly, in that case it can be reset to 0.
+			new LogUpdate(log => log.ParsingVersion < new Version(1, 11, 0, 2)
+			                     && log.Encounter == Encounter.TempleOfFebe,
+				"Add support for Temple of Febe Legendary CM"),
 		};
 
 		public IEnumerable<LogUpdateList> GetUpdates(IEnumerable<LogData> logs)