+++ ArcdpsLogManager/Sections/LogList.cs
@@ -165,6 +165,8 @@ namespace GW2Scratch.ArcdpsLogManager.Sections
 					{
 						switch (x.EncounterMode)
 						{
+							case EncounterMode.LegendaryChallenge:
+								return "LM";
 							case EncounterMode.Challenge:
 								return "CM";
 							case EncounterMode.Normal: